+++ apps/api/src/room/room.service.ts
@@ -318,10 +318,13 @@ export class RoomService {
      * @property {PositionType} payload.listeningSource - Current position of the listening source
      * @property {Record<string, { gain: number; rampTime: number }>} payload.gains - Gain configurations for each client
      */
-    const message: WsBroadcastType<{
+    
+    type BroadcastPayload = {
       listeningSource: PositionType;
       gains: Record<string, { gain: number; rampTime: number }>;
-    }> = {
+    };
+
+    const message: WsBroadcastType<BroadcastPayload> = {
       type: WsBroadcastTypeEnum.RoomMessage,
       payload: {
         listeningSource: room.listeningSource,
