+++ apps/web/.prettierrc
@@ -4,7 +4,7 @@
     "singleQuote": true,
     "tabWidth": 2,
     "trailingComma": "all",
-    "printWidth": 100,
+    "printWidth": 150,
     "bracketSpacing": true,
     "arrowParens": "avoid",
     "endOfLine": "lf"
