+++ package.json
@@ -16,6 +16,7 @@
     "@workspace/eslint-config": "workspace:*",
     "@workspace/typescript-config": "workspace:*",
     "prettier": "^3.5.1",
+    "prettier-plugin-organize-imports": "^4.1.0",
     "rimraf": "^6.0.1",
     "turbo": "^2.4.2",
     "typescript": "5.7.3"
